Proyectos IA /Pa09 /NewtonPy/projectFinal/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
openpyxl==3.1.2
scikit-learn==1.3.2
keras==2.15.0
tensorflow-macos==2.15.0
tensorflow-metal==1.2.0
nltk==3.8.1
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
scipy==1.11.4